add set_previous_message_id and set_message_id for DiffMessage, move tangle ref back to identity-iota
