+++ identity-iota-core/src/message/mod.rs
@@ -10,6 +10,5 @@ pub use iota_client::bee_message::Message;
 pub use iota_client::bee_message::MessageId;
 
 pub use self::traits::MessageIdExt;
-pub use self::traits::TangleRef;
 
 mod traits;
