+++ identity-iota/src/chain/diff_chain.rs
@@ -11,7 +11,6 @@ use identity_iota_core::diff::DiffMessage;
 use identity_iota_core::message::Message;
 use identity_iota_core::message::MessageId;
 use identity_iota_core::message::MessageIdExt;
-use identity_iota_core::message::TangleRef;
 use serde;
 use serde::Deserialize;
 use serde::Serialize;
@@ -25,6 +24,7 @@ use crate::tangle::Client;
 use crate::tangle::MessageExt;
 use crate::tangle::MessageIndex;
 use crate::tangle::PublishType;
+use crate::tangle::TangleRef;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[serde(transparent)]
