+++ identity-iota/src/diff/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod diff_message;
