+++ identity-iota/src/tangle/mod.rs
@@ -14,6 +14,7 @@ pub use self::publish::PublishType;
 pub use self::receipt::Receipt;
 pub use self::resolver::Resolver;
 pub use self::resolver::ResolverBuilder;
+pub use self::traits::TangleRef;
 pub use self::traits::TangleResolve;
 
 mod client;
