+++ identity-iota/src/tangle/traits.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use identity_iota_core::did::IotaDID;
+use identity_iota_core::message::MessageId;
 
 use crate::document::ResolvedIotaDocument;
 use crate::error::Result;
@@ -10,4 +11,16 @@ use crate::error::Result;
 #[async_trait::async_trait(?Send)]
 pub trait TangleResolve {
   async fn resolve(&self, did: &IotaDID) -> Result<ResolvedIotaDocument>;
+}
+
+pub trait TangleRef {
+  fn did(&self) -> &IotaDID;
+
+  fn message_id(&self) -> &MessageId;
+
+  fn set_message_id(&mut self, message_id: MessageId);
+
+  fn previous_message_id(&self) -> &MessageId;
+
+  fn set_previous_message_id(&mut self, message_id: MessageId);
 }
